feat: add EasyPost shipping provider integration
Implement complete EasyPost API integration for shipping labels:

- Add EasyPost provider with GetRates, CreateLabel, VoidLabel,
  TrackShipment, and ValidateAddress methods
- Add TenantID to all shipping params for multi-tenant security
- Store tenant_id in EasyPost shipment reference for validation
- Add idempotency check in CreateLabel (returns existing label)
- Change Rate.CostCents from int32 to int64 for future-proofing
- Add Rate.ExpiresAt to communicate 24-hour rate expiration
- Replace AddressValidation.IsValid bool with explicit status enum
- Enforce single-package shipments for MVP (multi-package post-MVP)
- Add structured logging with slog for observability
- Fix dollarsToCents parsing with proper decimal handling
- Use strings.Cut for cleaner rate ID parsing

+++ internal/service/checkout.go
@@ -205,7 +205,8 @@ func (s *checkoutService) CalculateOrderTotal(ctx context.Context, params OrderT
 		return nil, fmt.Errorf("failed to load cart: %w", err)
 	}
 
-	shippingCents := params.SelectedShippingRate.CostCents
+	// Convert int64 to int32 - safe for shipping costs which are always < $21M
+	shippingCents := int32(params.SelectedShippingRate.CostCents)
 
 	lineItems := make([]tax.LineItem, len(cartSummary.Items))
 	for i, item := range cartSummary.Items {
